Bugfixes and TS Info tab in AdminCP
Added some code to prevent a php notice which said that givenip is
undefined.
Added a TS Info tab which is accessed through the AdminCP -> Users &
Groups -> TeamSpeak Link -> TS Info tab.
Also the TABLE_PREFIX addition to the config file prevented the plugin
working from outside MyBB.
I left it in there but with an extra comment.