Content decoding fix and warning cleanup.
Fixed a possible content encoding failed error when using Use GZip Page
Compression.
Fixed a warning you would get when a connected TSUID wasn't found in the
TS database. (This can happen when a user connected a TS UID with his
MyBB user but it got pruned in the TS DB because the TS server prunes
his database according to the dbclientkeepdays setting in
ts3server.ini.)